Track wallet txs in OP and Base
